refactor: Change from customizaing hashing to disableHashing
