+++ packages/better-auth/src/plugins/api-key/index.ts
@@ -1,5 +1,3 @@
-import { base64Url } from "@better-auth/utils/base64";
-import { createHash } from "@better-auth/utils/hash";
 import { APIError, createAuthMiddleware } from "../../api";
 import type { BetterAuthPlugin } from "../../types/plugins";
 import { mergeSchema } from "../../db";
@@ -9,6 +7,7 @@ import { getDate } from "../../utils/date";
 import type { ApiKey, ApiKeyOptions } from "./types";
 import { createApiKeyRoutes } from "./routes";
 import type { User } from "../../types";
+import { defaultKeyHasher } from "./utilts";
 
 export const ERROR_CODES = {
 	INVALID_METADATA_TYPE: "metadata must be an object or undefined",
@@ -42,16 +41,6 @@ export const ERROR_CODES = {
 		"The property you're trying to set can only be set from the server auth instance only.",
 };
 
-const defaultKeyHasher = async (key: string) => {
-	const hash = await createHash("SHA-256").digest(
-		new TextEncoder().encode(key),
-	);
-	const hashed = base64Url.encode(new Uint8Array(hash), {
-		padding: false,
-	});
-	return hashed;
-};
-
 export const apiKey = (options?: ApiKeyOptions) => {
 	const opts = {
 		...options,
@@ -62,7 +51,7 @@ export const apiKey = (options?: ApiKeyOptions) => {
 		maximumNameLength: options?.maximumNameLength ?? 32,
 		minimumNameLength: options?.minimumNameLength ?? 1,
 		enableMetadata: options?.enableMetadata ?? false,
-		customKeyHasher: options?.customKeyHasher ?? defaultKeyHasher,
+		disableKeyHashing: options?.disableKeyHashing ?? false,
 		rateLimit: {
 			enabled:
 				options?.rateLimit?.enabled === undefined
@@ -159,7 +148,9 @@ export const apiKey = (options?: ApiKeyOptions) => {
 							});
 						}
 
-						const hashed = await opts.customKeyHasher(key);
+						const hashed = opts.disableKeyHashing
+							? key
+							: await defaultKeyHasher(key);
 
 						const apiKey = await ctx.context.adapter.findOne<ApiKey>({
 							model: schema.apikey.modelName,
