+++ packages/better-auth/src/plugins/api-key/utilts.ts
@@ -1,0 +1,12 @@
+import { base64Url } from "@better-auth/utils/base64";
+import { createHash } from "@better-auth/utils/hash";
+
+export const defaultKeyHasher = async (key: string) => {
+	const hash = await createHash("SHA-256").digest(
+		new TextEncoder().encode(key),
+	);
+	const hashed = base64Url.encode(new Uint8Array(hash), {
+		padding: false,
+	});
+	return hashed;
+};
